# openApi Documentation.yml
openapi: 3.1.0
x-stoplight:
  id: b83qtuy0a73jz
info:
  title: PilotesOrderAPI
  version: '1.0'
  summary: Api that allows the management of the pilotes orders
servers:
  - url: 'http://localhost:8080'
paths:
  /orders:
    parameters: []
    post:
      summary: Create Pilotes Order
      operationId: post-orders
      responses:
        '200':
          description: success
        '500':
          description: Internal Server Error
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                pilotes:
                  type: integer
                address:
                  type: object
                  properties:
                    street:
                      type: string
                    postcode:
                      type: integer
                    city:
                      type: string
                    country:
                      type: string
                client:
                  type: object
                  properties:
                    firstName:
                      type: string
                    lastName:
                      type: string
                    telephone:
                      type: integer
              x-examples:
                example-1:
                  pilotes: 5
                  address:
                    street: C Saturn 5 2-1
                    postcode: 25003
                    city: Lleida
                    country: Spain
                  client:
                    firstName: Alex
                    lastName: Mart
                    telephone: 600355441
            examples:
              example-1:
                value:
                  pilotes: 5
                  address:
                    street: C Saturn 5 2-1
                    postcode: 25003
                    city: Lleida
                    country: Spain
                  client:
                    firstName: Alex
                    lastName: Mart
                    telephone: 600355441
        description: |-
          {
              "pilotes": 5,
              "address": {
                  "street": "C Segovia 15 3-2",
                  "postcode": 25003,
                  "city": "Lleida",
                  "country": "Spain"
              },
              "client":{
                  "firstName": "Alex",
                  "lastName": "Mart",
                  "telephone": 600355441
              }
          }
      description: 'Creates a pilotes order with the client and address information. The amount of pilotes that can be orderes is only 5, 10 or 15.'
  '/orders/{id}':
    parameters:
      - schema:
          type: integer
        name: id
        in: path
        required: true
        description: Id of the pilotes order
    put:
      summary: ''
      operationId: put-user-id
      responses:
        '200':
          description: success
        '500':
          description: Internal Server Error
      description: 'Updates a pilotes order in the first 5 minutes following the creaton of the order. If the update happens after the first 5 minutes, it will give an error.'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                pilotes:
                  type: integer
                address:
                  type: object
                  properties:
                    street:
                      type: string
                    postcode:
                      type: integer
                    city:
                      type: string
                    country:
                      type: string
                client:
                  type: object
                  properties:
                    firstName:
                      type: string
                    lastName:
                      type: string
                    telephone:
                      type: integer
              x-examples:
                example-1:
                  pilotes: 5
                  address:
                    street: C Saturn 5 2-1
                    postcode: 25003
                    city: Lleida
                    country: Spain
                  client:
                    firstName: Alex
                    lastName: Mart
                    telephone: 600355441
            examples:
              example-1:
                value:
                  pilotes: 5
                  address:
                    street: C Saturn 5 2-1
                    postcode: 25003
                    city: Lleida
                    country: Spain
                  client:
                    firstName: Alex
                    lastName: Mart
                    telephone: 600355441
    get:
      summary: ''
      operationId: get-orders-id
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  client:
                    type: object
                    properties:
                      firstName:
                        type: string
                      lastName:
                        type: string
                      telephone:
                        type: integer
                x-examples:
                  example-1:
                    client:
                      firstName: Alex
                      lastName: Mart
                      telephone: 600355441
        '500':
          description: Internal Server Error
      description: Search orders by the client id.
components:
  schemas:
    PilotesOrder:
      type: object
      properties:
        pilotes:
          type: integer
        address:
          type: object
          properties:
            street:
              type: string
            postcode:
              type: integer
            city:
              type: string
            country:
              type: string
        client:
          type: object
          properties:
            firstName:
              type: string
            lastName:
              type: string
            telephone:
              type: integer
      x-examples:
        example-1:
          pilotes: 5
          address:
            street: C Saturn 5 2-1
            postcode: 25003
            city: Lleida
            country: Spain
          client:
            firstName: Alex
            lastName: Mart
            telephone: 600355441
      title: PilotesOrder
      description: Represents the Pilotes Order
    Client:
      type: object
      x-examples:
        example-1:
          client:
            firstName: Alex
            lastName: Mart
            telephone: 600355441
      properties:
        client:
          type: object
          properties:
            firstName:
              type: string
            lastName:
              type: string
            telephone:
              type: integer
      title: Client
      description: Represents the client information
    Address:
      type: object
      properties:
        address:
          type: object
          properties:
            street:
              type: string
            postcode:
              type: integer
            city:
              type: string
            country:
              type: string
      x-examples:
        example-1:
          address:
            street: C Saturn 5 2-1
            postcode: 25003
            city: Lleida
            country: Spain
      description: Represents the delivery address of the order
      title: Address